app/requirements.txt
fastapi
uvicorn[standard]
pillow
python-multipart
numpy
tensorflow